+++ range/Range/src/gl_scalar_field.cpp
@@ -210,7 +210,7 @@ std::vector<ScalarFieldItem> GLScalarField::calculateField(const RVariable *pSca
     double maxValue = rVariable.getVariableData().getMaxDisplayValue();
 
 #pragma omp parallel for default(shared)
-    for (uint i=0;i<this->elementGroupIDs.size();i++)
+    for (int64_t i=0;i<int64_t(this->elementGroupIDs.size());i++)
     {
         REntityGroupType entityType;
         uint entityID;
