+++ range/RangeBase/include/rbl_ivector.h
@@ -39,7 +39,7 @@ class RIVector : public std::vector<int>
         //! Return number of rows (size of the vector).
         inline unsigned int getNRows ( void ) const
         {
-            return this->size();
+            return (unsigned int)this->size();
         }
 
         //! Access element operator.
